// src/types/my-purchase.ts
export interface MyPurchase {
    listing_id: number;
    book_title: string;
    price: number;
    transaction_time: string; 
    image_url: string;
    seller_first_name: string;
    seller_last_name: string;
    seller_profile_img: string;
    seller_phone: string;
    book_id: number;
    seller_id: number;
  }